Started on creating a branch for removing the store locator functionality which belongs in a separate module.
This code is entirely tested and should not be used in its current state.

Changes were made to the installer and upgrade scripts in order to avoid creating unnecessary tables, and so that we can remove an unused table from config.xml. This isn't ideal, however, it was the cleanest of the options available. The installers do not need to be rerun, they only eliminate a table and fields that are no longer used and won't cause any issues (that we're aware of) if they remain.