Cleanup the modman file
There isn't much reason to manually specify every single file in the module to be linked individually.
This actually harms development as new files will not be seen until listed in modman and linked.

Additionally, the second column can/should be left out as modman supports not providing that information
if the destination matches the source. The only reason no to do this is the Magento composer installer
not properly handling modman files.